Week1-Task5: Add a second, read-only component that shows $BTC instead of Euros; use this function to get the exchange rate:
